correcciones

+++ AdvanceApi/Clases/Operacion.cs
@@ -7,9 +7,9 @@ namespace Clases
         public int IdOperacion { get; set; }
         public string? Concepto { get; set; }
         public int? IdCliente { get; set; }
-        public decimal? Monto { get; set; }
-        public decimal? Abono { get; set; }
-        public decimal? Restante { get; set; }
+        public double? Monto { get; set; }
+        public double? Abono { get; set; }
+        public double? Restante { get; set; }
         public string? Nota { get; set; }
         public bool? Estatus { get; set; }
         public DateTime? FechaInicio { get; set; }
